Create start and end steps with fake offline syncing
I have left the offline tests commented out as we don't currently have a
good way to control the network of the running browser. We also need to
consider how this should work for server side rendering, though perhaps
if it came from the server, we should already know they're online?