Persist the data fetched from the API to storage if it's newer
We don't want to accidentally overwrite offline data with the data from
the API, as it might be older, if they haven't synced yet.

In order for this to work, we now update the last modified time when we
change steps via the `Orchstrator`. It would be better to be able to do
at as an automatic step after any database update, but this does the job
for now.