Clear the process data after submission
We don't actually send it anywhere, but when we do, we'll want to clear
out the local copy so we'll never end up with accidental edits to the
local copy conflicting with the stored copy. We also have a limited
amount of space, so we want to clear out old data we don't need.