Switch to using a base path over a parameter to support filtering
The load balancer needs to know which process a request belongs to. We
were using a query parameter, but that's messy and doesn't work for
asset requests.

Now, we use a reverse proxy via a custom server, and Next's Link's
`href` vs `as` features to map `/thc/*` to `/*`, supporting filtering by
URL pattern.

This requires enabling experimental catch-all routing for Next, which
isn't covered by semver. We need this to support `as="/thc/*"` mapping
to `href="/[...slug]"`, which otherwise wouldn't match.