Move the pages to expect a process ref URL prefix
We want to move away from using session storage to track the current
process, as that's fragile when working on two processes simultaneously.
To do this, we think adding the process ref to the URL is more reliable.

The router's only available on the client, so we need to pass it in from
a safe source (like `useRouter`) rather than using the singleton.

This reverts the previous change to use the experimental base path
setting. It doesn't work well with dynamic paths. Instead, we now
enclose everything in a `thc` directory instead. This is more reliable
and predictable.