Add rendeing of components
Added the is maintained radio buttons if choice is "not -sure" for garden type. Also added notes box if garden is not maintained.
The notes box is already there if the choice is maintained but appears below the take photo button. May need to play around with the positioning of these as wasn't sure how to implament && or || without repeating a renderWhen for the no condition.