Update the resident database at the same time as the process database
We do it before the process database update, so `lastModified` only gets
updated after everything else, to avoid data loss in the case of errors.
We run the migrations if the incoming data has a later last modified
date, so changing it means future updates won't run.

We also now remove data not present in the incoming process data. We
don't want stale data sneaking back in to processes.