Explicitly specify inherited `renderWhen` conditions
Without this change, filling in a chain of questions then returning to
an earlier question in the chain and changing the answer results in the
questions that depend on questions no longer visible remain visible.
This is because the no invisible question's answer is still stored ready
to be submitted. We do this because we don't want to destroy the user's
answers due to e.g. a misclick.